requirements.txt
streamlit
numpy
matplotlib
POT
requests
pillow
